Work around ARM and IAR optimization issue
Add a custom profile to work around an issue with the ARM and IAR
compilers, where the example appears to hang without any error. Building
the example with `-O1` (ARM) and `-Ol` (IAR), rather than the default of
`-Os` (ARM) and `-Ohz` (IAR), works as a workaround.

All toolchains, not just ARM and IAR, are added to the profile in the
case where the user wishes to always specify the --profile selecting
options even when using a toolchain which doesn't exhibit faulty
behavior.

To use the custom profile with mbed-cli, along with the develop profile,
use the --profile option to provide the develop profile followed by the
custom override-only profile.

    mbed compile -m CY8CKIT_062_WIFI_BT -t ARMC6 --profile mbed-os/tools/profiles/develop.json --profile mbed_profile.json

Arm internal reference: SDCTRESPONSE-3550