Include "mbed.h" to enable GCC_ARM build with "-flto"
Add the "mbed.h" header to use the MBED_USED attribute with main().

Since the "used" attribute and the definition of main() have to be in
the same translation unit, I placed a declaration of main() in
"platform/mbed_toolchain.h" which "mbed.h" includes.

Without the attribute, the main() symbol is not emitted with the GCC
toolchain using "-Wl,--wrap,main" and "-flto" flags.