Remove PSA Crypto freeing and refactor initialization
The function `mbedtls_psa_crypto_free()` is not part of the PSA Crypto
API and not provided by TF-M, so we avoid calling it in the
application. Consequently it does not make sense anymore to call
`psa_crypto_init()` repeatedly in all the examples in the application,
so we call it once at the start of `main()` only.