getting-started: Fix PSA macro usage
The latest version of TF-M (1.4.0) uses the newer PSA API. In the later
PSA API some of the macros we make use of in this example have been
renamed. We still have some targets that make use of the older version
of TF-M and PSA, so we need to keep backwards compatibility for these
targets.